refactor: move components in accordance with atomic design
