+++ src/pages/users/[id].tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import Container from '../../components/atoms/Container';
 import NavBar from '../../components/molecules/NavBar';
-import TextList from '../../components/TextList';
+import TextList from '../../components/organisms/text/TextList';
 import UserAvatar from '../../components/atoms/UserAvatar';
 import UserName from '../../components/organisms/UserName';
 import useUser from '../../hooks/useUser';
